Pull out drag and drop components into module
